первичная попытка оптимизации
